weatherapp/requirements.txt
requests
flask
flask-mysqldb
boto3